fix: Add CRUD of members and user registration with generating username and password

+++ app/Repositories/Eloquent/SocioRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories\Eloquent;
 
 use App\Models\Socio;
 use App\Repositories\Interfaces\SocioRepositoryInterface;
+use Illuminate\Support\Facades\DB;
 
 class SocioRepository implements SocioRepositoryInterface {
     // Lista de todos los socios con Eloquent
@@ -21,6 +22,7 @@ class SocioRepository implements SocioRepositoryInterface {
     // Registro de un nuevo socio
     public function create(array $data)
     {
+        DB::beginTransaction();
         $socio = Socio::create([
             'nombre_socio' => $data['nombre'],
             'primer_apellido_socio' => $data['primer_apellido'],
@@ -31,20 +33,36 @@ class SocioRepository implements SocioRepositoryInterface {
         ]);
 
         if($socio){
-            return 'Registro del socio exitoso';
+            DB::commit();
+            return $socio;
         }
-        return 'Error al crear el socio.';
+        DB::rollBack();
+        return null;
     }
 
     public function update(array $data, string $id) {
-        $socio = Socio::find($id);
-        $socio->update($data);
+
+        $socio = Socio::find(1);
+        if(!$socio){
+            return null;
+        }
+        $socio->update([
+            'nombre_socio' => $data['nombre'],
+            'primer_apellido_socio' => $data['primer_apellido'],
+            'segundo_apellido_socio' => $data['segundo_apellido'],
+            'ci_socio' => $data['ci'],
+            'image' => $data['image'],
+            'otb_id' => 1
+        ]);
         return $socio;
     }
 
     public function delete(string $id) {
         $socio = Socio::find($id);
+        if(!$socio){
+            return 'Socio no encontrado.';
+        }
         $socio->delete();
-        return $socio;
+        return 'Socio eliminado correctamente.';
     }
 }
